requirements.txt
pysnmp
paho-mqtt